implement authorization using gates

+++ app/Http/Controllers/QuestionController.php
@@ -63,8 +63,10 @@ class QuestionController extends Controller
      */
     public function edit(Question $question)
     {
-        
-        return view("questions.edit", compact('question'));
+        if(\Gate::denies('update-question', $question)){
+            abort(403,"Access denied");
+        }
+        return view("questions.edit", compact('question'));  
     }
 
     /**
@@ -88,7 +90,13 @@ class QuestionController extends Controller
      */
     public function destroy(Question $question)
     {
+        if(\Gate::denies('delete-question', $question)){
+            abort(403,"Access denied");
+        }
         $question->delete();
         return redirect('/questions')->with('success',"Your question has been deleted");
     }
 }
+
+
+
